Implements a new download feature and pushes a new GitHub Action workflow.
