+++ .github/workflows/on_push.yml
@@ -1,0 +1,111 @@
+name: Validation
+
+on: [push, pull_request]
+  push:
+    branches:
+      - master
+  pull_request:
+
+jobs:
+  runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [windows-2019, ubuntu-16.04, ubuntu-18.04, macos-10.15]
+        toolchain: ["stable", "beta", "nightly", "1.32.0"]
+  build:
+    name: Builds the project
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Install the Rust toolchain
+          uses: actions-rs/toolchain@v1
+          with:
+            toolchain: ${{ matrix.toolchain }}
+            override: true
+            default: true
+
+      - name: Attempt to restore from cache
+        uses: actions/cache@v1
+        with:
+          path: target
+          key: ${{ matrix.os }}-${{ rustc }}-build-${{ hashFiles('Cargo.lock') }}
+
+      - name: Build
+        id: build
+        run: cargo build --debug --verbose
+
+      - name: Upload build
+          uses: actions/upload-artifact@v1
+          with:
+            name: build
+            path: target
+
+  conventions:
+    name: Builds the project
+    needs: build
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Install the Rust toolchain
+          uses: actions-rs/toolchain@v1
+          with:
+            toolchain: ${{ matrix.toolchain }}
+            components: rustfmt, clippy
+            override: true
+            default: true
+
+      - name: Attempt to restore from cache
+        uses: actions/cache@v1
+        with:
+          path: target
+          key: ${{ matrix.os }}-${{ rustc }}-build-${{ hashFiles('Cargo.lock') }}
+
+      - name: Perform unit testing and integration testing
+        run: cargo test --workspace --verbose
+
+  testing:
+    name: Builds the project
+    needs: build
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Install the Rust toolchain
+          uses: actions-rs/toolchain@v1
+          with:
+            toolchain: ${{ matrix.toolchain }}
+            components: rustfmt, clippy
+            override: true
+            default: true
+
+      - name: Attempt to restore the compiled dependencies from cache
+        uses: actions/cache@v1
+        with:
+          path: target
+          key: ${{ matrix.os }}-${{ rustc }}-build-${{ hashFiles('Cargo.lock') }}
+
+      - name: Perform unit testing and integration testing
+        run: cargo test  --workspace --all-targets --verbose
+
+  documentation-tests:
+    name: Builds the project
+    needs: build
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Install the Rust toolchain
+          uses: actions-rs/toolchain@v1
+          with:
+            toolchain: ${{ matrix.toolchain }}
+            components: rustfmt, clippy
+            override: true
+            default: true
+
+      - name: Attempt to restore the compiled dependencies from cache
+        uses: actions/cache@v1
+        with:
+          path: target
+          key: ${{ matrix.os }}-${{ rustc }}-build-${{ hashFiles('Cargo.lock') }}
+
+      - name: Perform unit testing and integration testing
+        run: cargo test --doc --verbose
+
