+++ .github/workflows/on_tag.yml
@@ -1,0 +1,17 @@
+name: Validation
+
+on:
+  push:
+    tags:
+      - v*
+
+jobs:
+  build:
+    name: Builds the project
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Build
+        run: cargo build --verbose
+      - name: Run tests
+        run: cargo test --verbose
