--- .travis.yml
@@ -1,32 +1,0 @@
-language: rust
-cache: cargo
-sudo: required
-
-rust:
-  - stable
-  - beta
-
-before_script:
-  - rustup component add clippy
-
-addons:
-  apt:
-    packages:
-      - libcurl4-openssl-dev
-      - libelf-dev
-      - libdw-dev
-      - cmake
-      - gcc
-      - binutils-dev
-      - libiberty-dev
-
-script:
-  - cargo build --release --verbose
-  - cargo test --release --verbose
-  - cargo doc --release --verbose
-  - cargo clippy --release --verbose
-
-after_success:
-  - wget https://github.com/SimonKagstrom/kcov/archive/master.tar.gz && tar xzf master.tar.gz && cd kcov-master && mkdir build && cd build && cmake .. && make && make install DESTDIR=../../kcov-build && cd ../.. && rm -rf kcov-master
-  - for file in target/debug/*[^\.d]; do mkdir -p "target/cov/$(basename $file)"; ./kcov-build/usr/local/bin/kcov --exclude-pattern=/.cargo,/usr/lib --verify "target/cov/$(basename $file)" "$file"; done;
-  - bash <(curl -s https://codecov.io/bash)
