+++ Cargo.toml
@@ -1,10 +1,10 @@
 [package]
 name = "rsef-rs"
-version = "0.1.3"
+version = "0.2.0"
 authors = ["Christian Veenman <chris_veenman@hotmail.com>"]
 edition = '2018'
 readme = "README.md"
-keywords = ["rsef", "parser"]
+keywords = ["rsef", "parser", "rir", "ripe", "apnic", "lacnic", "arin", "afrinic"]
 categories = ["parsing"]
 repository = "https://github.com/DevQps/rsef-rs"
 homepage = "https://github.com/DevQps/rsef-rs"
@@ -14,16 +14,22 @@ license = "GPL-3.0"
 exclude = [
     "README.md",
     "tests/*",
-    ".travis.yml"
+    ".github/*"
 ]
 
 [badges]
-travis-ci = { repository = "DevQps/rsef-rs", branch = "master" }
 codecov = { repository = "DevQps/rsef-rs", branch = "master", service = "github" }
 maintenance = { status = "actively-developed" }
 
+[features]
+# Allows a user to download the RSEF listings.
+download = ["reqwest", "bzip2", "libflate", "chrono"]
+
+# No feature is included in the default distribution.
+default = []
+
 [dependencies]
-reqwest = "0.9.9"
-bzip2 = "0.3.3"
-libflate = "0.1"
-chrono = "0.4.6"
+reqwest = { version = "0.9", optional = true }
+bzip2 = { version = "0.3", optional = true }
+libflate = { version = "1.0", optional = true }
+chrono = { version = "0.4", optional = true }
